feat(web): integrate @og-tester/mcp server as Next.js API route via xmcp adapter
