+++ apps/web/app/api/mcp/route.ts
@@ -1,0 +1,3 @@
+import { xmcpHandler } from "@xmcp/adapter";
+
+export { xmcpHandler as GET, xmcpHandler as POST };
