--- packages/mcp/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@og-tester/mcp",
-  "version": "0.0.0",
-  "bin": {
-    "og-tester-mcp": "./dist/stdio.js"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "dev": "xmcp dev",
-    "build": "xmcp build",
-    "start": "node dist/stdio.js",
-    "typecheck": "tsgo --noEmit"
-  },
-  "dependencies": {
-    "@og-tester/core": "workspace:*",
-    "xmcp": "latest",
-    "zod": "^4.3.6"
-  },
-  "devDependencies": {
-    "@og-tester/typescript-config": "workspace:*",
-    "typescript": "^5.9.3"
-  }
-}
